# src/main/resources/Actor and Movie API swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Actor and Movie API
host: unknown
schemes:
  - https
  - http
basePath: /v1
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
paths:
  /actors:
    get:
      summary: Retrieve a list of actors
      description: >
        The list is ordered by actor name, ascending.

        Paging is done by using parameters page and page_size.

        Max page_size is 1000.


        A global (common for all endpoints) rate limit of 5 requests per minute
        and user account applies.
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: number
          required: false
          default: 0
        - name: page_size
          in: query
          type: number
          required: false
          default: 10
        - name: name
          in: query
          type: string
          required: false
          description: Filter by actor name (wildcard search)
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/actor'
        '401':
          description: Unauthorized
        '429':
          description: Too many requests. Rate limit exceeded.
  '/actors/{id}':
    get:
      summary: Retrieve an actor
      description: >
        A global (common for all endpoints) rate limit of 5 requests per minute
        and user account applies.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: number
          required: true
          description: Actor ID
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/actor'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '429':
          description: Too many requests. Rate limit exceeded.
  '/actors/{id}/appearances':
    get:
      summary: Retrieve a list of appearances for an actor
      description: >
        The list is ordered by movie name, ascending.

        Paging is done by using parameters page and page_size.

        Max page_size is 1000.


        A global (common for all endpoints) rate limit of 5 requests per minute
        and user account applies.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: number
          required: true
          description: Actor ID
        - name: page
          in: query
          type: number
          required: false
          default: 0
        - name: page_size
          in: query
          type: number
          required: false
          default: 10
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/appearance'
        '401':
          description: Unauthorized
        '429':
          description: Too many requests. Rate limit exceeded.
  /movies:
    get:
      summary: Retrieve a list of movies
      description: >
        The list is ordered by movie name, ascending.

        Paging is done by using parameters page and page_size.

        Max page_size is 1000.


        A global (common for all endpoints) rate limit of 5 requests per minute
        and user account applies.
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: number
          required: false
          default: 0
        - name: page_size
          in: query
          type: number
          required: false
          default: 10
        - name: name
          in: query
          type: string
          required: false
          description: Filter by movie name (wildcard search)
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/movie'
        '401':
          description: Unauthorized
        '429':
          description: Too many requests. Rate limit exceeded.
  '/movies/{id}':
    get:
      summary: Retrieve a movie
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: number
          required: true
          description: Movie ID
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/movie'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '429':
          description: Too many requests. Rate limit exceeded.
definitions:
  actor:
    properties:
      id:
        type: number
      name:
        type: string
  movie:
    properties:
      id:
        type: number
      title:
        type: string
      year:
        type: string
  appearance:
    properties:
      movie_id:
        type: number
      movie_name:
        type: string
      character_name:
        type: string